Changed window retry as 3 times
Changed window retry as 3 times, this will improve the wait duration

Compared to new changes (recent fixes) [1] I see some improvement (with my today's fix) [2], but not equivalent to earlier cobra version [3]. In my test run [2], BAT2 and 3 with # 1 and 2 in the result set are with retry as 3 and # 3 and 4 are with retry as 2. Based on this, I'm leaving the code with 3 retries. Note: I have removed delay of 1 second between each retry.

[1] - http://racetrack-dev.eng.vmware.com/result.php?id=345238
[2] - http://racetrack-dev.eng.vmware.com/result.php?id=345393
[3] - http://racetrack.eng.vmware.com/result.php?id=963433

QA Notes:
Testing Done: VMware Workstation BATS
Documentation Notes:
Bug Number:
Reviewed by:
Approved by:
Mailto: